Restore network jungle awareness

+++ 03_voxp_domain/src/main/java/ru/voxp/android/domain/api/model/ResponseContainer.kt
@@ -1,7 +1,7 @@
 package ru.voxp.android.domain.api.model
 
 class ResponseContainer {
-    var count: Long? = null
+    var count: Int? = null
     var page: Int? = null
     var laws: List<Law>? = null
 }
